chore: move fullscreen handling to a hook
